requirements.txt
asgiref==3.5.2
aws-xray-sdk==2.10.0
boto3==1.26.2
botocore==1.29.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-environ==0.9.0
django-extensions==3.2.1
django-health-check==3.17.0
django-storages==1.13.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
ecdsa==0.18.0
idna==3.4
jmespath==1.0.1
psycopg2-binary==2.9.5
pyasn1==0.4.8
PyJWT==2.6.0
python-dateutil==2.8.2
python-jose==3.3.0
python-keycloak==2.6.0
pytz==2022.6
requests==2.28.1
requests-toolbelt==0.9.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
watchtower==3.0.0
wrapt==1.14.1
